Web server on sockets
Develop a server program that accepts a browser request as input. The web server parses the start string of the request to determine which file to return in response to the request. The file is identified by the URL path. The response will be textual information, an HTTP response, which contains the response headers (Server, Date, Content-Type, Connection, Content-Encoding) and file content.
The program should return 200, 404 and 5xx responses. For a single URL path, you must throw a 302 or 301 error.
Launch the server program and display files in the browser: Text, HTML, JSON, JPG and MP4. The results of the answers should be included in the report.
Parse JSON file request and response using Wireshark sniffer. The results of the analysis should be included in the report.
Modify the server program. When you first access the web server, prompt the browser to save cookies that contain the user id. On each connection without passing a cookie, increase id. On receipt of the cookie, allow the URL path to navigate to the "secret" JSON file.
Report cookies sent from the server and received from the client using Wireshark.
For the Google Chrome browser, do not output files with the MP4 extension. Give an example of a file opening from another browser.
The code of the final version of the program should be included in the report.